Consolidate hardware info all into one page
Info that's not relevant to users or shouldn't necessarily be public have been
removed.

The networking info isn't a big deal, but it's best not to advertise the info,
and has been moved to internal documents.

Much of this info is changing soon anyway, with services being migrated,
containerized, and deployed using ansible. And the networking changes necessary
to go along with it.