IPython problems on the headnode.
The issue is IPython history, which uses SQLite, and that our home folders are
shared over NFS.

https://ipython.readthedocs.io/en/stable/config/options/terminal.html#configtrait-HistoryAccessor.hist_file

Except... that's not the problem. This has been working for years.

The updates that triggered this problem are:
  libpolkit-agent-1-0 [0.105-18 -> 0.105-18+deb9u1]
  libpolkit-backend-1-0 [0.105-18 -> 0.105-18+deb9u1]
  libpolkit-gobject-1-0 [0.105-18 -> 0.105-18+deb9u1]
  policykit-1 [0.105-18 -> 0.105-18+deb9u1]

This is reproducible on the compute nodes (which didn't have these updates).
Applying the update causes the problem. Rebooting the node clears the problem.

No known contortion of mounting, unmounting, etc resolves the problem.

Upgrading a node, getting the problem, and then issuing ``service polkit
restart`` resolves the problem. But not on the head node.

I am considering murdering kittens at the moment.